requirements.txt
amqp==5.0.6
asgiref==3.2.10
billiard==3.6.4.0
celery==5.1.2
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==3.1
django-admin-index==1.5.0
django-filter==2.4.0
django-ordered-model==3.4.3
djangorestframework==3.12.4
kombu==5.1.0
numpy==1.21.2
pandas==1.3.3
Pillow==8.3.2
prompt-toolkit==3.0.20
Pygments==2.10.0
python-dateutil==2.8.2
pytz==2021.1
redis==3.5.3
six==1.16.0
sqlparse==0.4.2
vine==5.0.0
wcwidth==0.2.5
